Refactor write. Now the writing of the data will be done in the addEntry method
request-checks: true

+++ Noisrev.League.IO.RST/Helper/OffsetHelper.cs
@@ -16,7 +16,7 @@ namespace Noisrev.League.IO.RST.Helper
         /// <exception cref="OverflowException"/>
         public static ulong ComputeOffset(this long offset, RType type)
         {
-            return Convert.ToUInt64(offset << (int)(type));
+            return Convert.ToUInt64(offset << (byte)type);
         }
     }
 }
